explicitly specify int

+++ Dinic.cpp
@@ -11,8 +11,8 @@ vi iter(MAX_V);       // どこまで調べ終わったか
 
 // fromからtoへ向かう容量capの辺をグラフに追加する
 void add_edge(int from, int to, lli cap){
-  G[from].push_back((edge){to, cap, G[to].size()});
-  G[to].push_back((edge){from, 0, G[from].size() - 1});
+  G[from].push_back((edge){to, cap, (int)G[to].size()});
+  G[to].push_back((edge){from, 0, (int)G[from].size() - 1});
 }
 
 // sからの最短距離をBFSで計算する
